지도 여러번 그려지는 작업 수정 및 페이지 리팩토링
