+++ src/components/FavoriteBuildings.js
@@ -18,7 +18,7 @@ export default function FavoriteBuildings() {
 
       setFavoriteBuildings(buildingArray);
     })();
-  }, [favoriteBuildings]);
+  }, []);
 
   const showBuildingDetailPage = (event) => {
     navigate(`/detail/${event.target.id}`);
@@ -26,13 +26,10 @@ export default function FavoriteBuildings() {
 
   const deleteFavoriteBuilding = async (event) => {
     const id = event.target.id;
-    const newBuildingsArray = [];
 
-    for (let i = 0; i < favoriteBuildings.length; i++) {
-      if (favoriteBuildings[i]._id !== id) {
-        newBuildingsArray.push(favoriteBuildings[i]);
-      }
-    }
+    const newBuildingsArray = favoriteBuildings.filter(
+      (building) => building._id !== id,
+    );
 
     const user = await useAxios(
       `/users/user/favorites/buildings/${id}`,
