+++ src/hooks/useMap.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { getGraphData } from '../utils/graph';
-import { debounce } from 'lodash';
+import _ from 'lodash';
 
 import useAxios from './useAxios';
 
@@ -11,6 +11,7 @@ export default function useMap(position, type, mapElement) {
   const [currentAddress, setCurrentAddress] = useState('');
   const [currentCenter, setCurrentCenter] = useState([37.5, 127.0]);
   const [graphData, setGraphData] = useState({});
+  const [isMap, setIsMap] = useState(false);
   const infoWindowArray = [];
   const auctionMarkers = [];
   const forSalesArray = [];
@@ -21,6 +22,7 @@ export default function useMap(position, type, mapElement) {
 
   useEffect(() => {
     const mapContainer = mapElement.current;
+
     const mapOptions = {
       center: new kakao.maps.LatLng(currentCenter[0], currentCenter[1]),
       level: 2,
@@ -28,7 +30,7 @@ export default function useMap(position, type, mapElement) {
       scrollwheel: true,
     };
 
-    const map = new kakao.maps.Map(mapContainer, mapOptions);
+    const map = !isMap ? new kakao.maps.Map(mapContainer, mapOptions) : isMap;
     const geocoder = new kakao.maps.services.Geocoder();
     const ps = new kakao.maps.services.Places();
 
@@ -39,6 +41,7 @@ export default function useMap(position, type, mapElement) {
     });
 
     cluster = clusterer;
+
     if (position) {
       const coord = new kakao.maps.LatLng(position);
 
@@ -74,7 +77,8 @@ export default function useMap(position, type, mapElement) {
       deleteForSaleMarkers(map);
     }
 
-    mapContainer.addEventListener('touchend', getMaxDistance(map, geocoder));
+    setIsMap(map);
+    mapContainer.ontouchend = getMaxDistance(map, geocoder);
   }, [place, type, showAll]);
 
   const getMaxDistance = (map, geocoder) => {
