--- src/pages/Alarm.js
@@ -1,10 +1,0 @@
-import NavBar from '../components/NavBar';
-
-export default function Alarm() {
-  return (
-    <>
-      <button></button>
-      <NavBar />
-    </>
-  );
-}
