+++ src/routes/Authorized.js
@@ -3,7 +3,6 @@ import { Routes, Route, Navigate } from 'react-router-dom';
 import Detail from '../pages/Detail';
 import Main from '../pages/Main';
 import SearchPlace from '../pages/SearchPlace';
-import Alarm from '../pages/Alarm';
 import MyPage from '../pages/MyPage';
 import MyFavoriteRegion from '../pages/FavoriteRegion';
 
@@ -15,7 +14,6 @@ export default function Authorized() {
       <Route path='/detail/:buildingId' element={<Detail />} />
       <Route path='/search' element={<SearchPlace />} />
       <Route path='/search/:place' element={<SearchPlace />} />
-      <Route path='/alarm' element={<Alarm />} />
       <Route path='/mypage' element={<MyPage />} />
       <Route path='/favoriteregion' element={<MyFavoriteRegion />} />
     </Routes>
